Moved firebase.ts into src and updated routes referencing it
