+++ functions/.eslintrc.js
