+++ functions/src/userDeleteFunction.ts
