+++ src/types/Blacklisted.ts
